% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePPP.R
\name{calculatePPP}
\alias{calculatePPP}
\title{Compute the posterior predictive p}
\usage{
calculatePPP(data, origMCMCSamples, ndraws, sampleFromPP, discrepancy, ...)
}
\arguments{
\item{data:}{original data}

\item{origMCMCSamples:}{matrix of posterior samples}

\item{ndraws:}{number of samples to simulate}

\item{sampleFromPP:}{function that can sample new data from the posterior predictive using the provided samples}

\item{discrepancy:}{a function that calculates the discrepancy (statistics) for the data (original or sampled)}

\item{...:}{additional arguments to be passed}
}
\value{
list containing the p-value, sampled discrepancies and original one
}
\description{
function that simulate samples from the posterior predictive and calculate summaries
}
