% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{computeCppp}
\alias{computeCppp}
\title{Compute calibrated posterior predictive p-value (placeholder)}
\usage{
computeCppp(pppObs, pppRep)
}
\arguments{
\item{pppObs}{Numeric scalar in \eqn{[0,1]}; observed PPP.}

\item{pppRep}{Numeric vector in \eqn{[0,1]}; calibration PPPs.}
}
\value{
Numeric scalar; calibrated PPP.
}
\description{
Compute calibrated posterior predictive p-value (placeholder)
}
