% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpppPrototype-package.R
\docType{package}
\name{cpppPrototype-package}
\alias{cpppPrototype}
\alias{cpppPrototype-package}
\title{cpppPrototype: Posterior Predictive Checks and Calibrated P-values}
\description{
Tools for performing posterior predictive checks and computing calibrated posterior predictive p-values (cPPP) in Bayesian models. Includes infrastructure for calibration loops, diagnostic summaries, and NIMBLE-based examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/salleuska/cpppPrototype}
  \item Report bugs at \url{https://github.com/salleuska/cpppPrototype/issues}
}

}
\author{
\strong{Maintainer}: Sally Paganin \email{paganin.1@osu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Perry de Valpine (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Daniel Turek (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
