% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpppResult-class.R
\name{new_cppresults}
\alias{new_cppresults}
\title{Create a cpppResult object}
\usage{
new_cppresults(
  cppp = NA_real_,
  ppp = numeric(),
  obs_ppp = numeric(),
  discrepancies = NULL,
  ...
)
}
\arguments{
\item{cppp}{Numeric scalar. The approximated calibrated posterior predictive p-value.}

\item{ppp}{Numeric vector. The posterior predictive p-values across calibration datasets.}

\item{obs_ppp}{Numeric vector. The observed posterior predictive p-values for the observed data.}

\item{discrepancies}{Optional list or array containing discrepancy values for
each replication and dataset (observed and simulated).}

\item{...}{Additional elements to store in the object.
SP E.g., do we want to save in this object functions used in the pipelie?}
}
\value{
An object of class \code{cppresults}.
}
\description{
Constructs an S3 object of class \code{cpppResult}, which stores the results
related to the calibration of the predictive p-value.
}
