% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibrationNIMBLE.R
\name{runCalibrationNIMBLE}
\alias{runCalibrationNIMBLE}
\title{Run calibration using a NIMBLE model}
\usage{
runCalibrationNIMBLE(
  model,
  dataNames = NULL,
  paramNames,
  disc_fun,
  n_reps = 100,
  MCMCcontrolMain = list(niter = 5000, nburnin = 1000, thin = 1),
  MCMCcontrolRep = list(niter = 500, nburnin = 0, thin = 1),
  mcmcConfFun = NULL,
  row_selector = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{model}{Compiled nimbleModel with observed data set.}

\item{dataNames}{Optional character vector of data node names. If NULL,
nodes flagged as data in the model are used.}

\item{paramNames}{Character vector of parameter node names to monitor.}

\item{disc_fun}{Function \verb{function(data, theta_row, control)} returning a
scalar or vector discrepancy for one posterior draw.}

\item{n_reps}{Number of calibration replications.}

\item{MCMCcontrolMain}{List with \code{niter}, \code{nburnin}, \code{thin} for main chain.}

\item{MCMCcontrolRep}{List with \code{niter}, \code{nburnin}, \code{thin} for calibration chains.}

\item{mcmcConfFun}{Optional function \verb{function(model)} returning an MCMC configuration.}

\item{row_selector}{Optional row selector (see runCalibration()).}

\item{control}{List of additional options passed to discrepancy / helpers.}

\item{...}{Not used currently.}
}
\description{
Run calibration using a NIMBLE model
}
