% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunctions.R
\name{computeCppp}
\alias{computeCppp}
\title{Compute the calibrated posterior predictive p-value}
\usage{
computeCppp(pppObs, pppCal)
}
\arguments{
\item{pppObs}{Numeric scalar in \eqn{[0,1]}. The observed PPP.}

\item{pppCal}{Numeric vector in \eqn{[0,1]}. PPPs from calibration datasets.}
}
\value{
Numeric scalar in \eqn{[0,1]}, the calibrated PPP.
}
\description{
Compute the calibrated posterior predictive p-value
}
