% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{runCalibration}
\alias{runCalibration}
\title{Run calibration}
\usage{
runCalibration(
  MCMC_samples,
  MCMC_fun,
  new_data_fun,
  disc_fun,
  num_reps,
  control = NULL,
  ...
)
}
\arguments{
\item{MCMC_samples}{A matrix or list containing posterior samples from the abserved-data (long) run. Passed to \code{new_data_fun()} as context.}

\item{MCMC_fun}{A function of the form \verb{function(new_data, control, ...)} that runs an MCMC algorithm and returns posterior samples.}

\item{new_data_fun}{A function of the form \verb{function(MCMC_samples, ...)} that generates a new synthetic dataset for each calibration replicate. SP: We assume that new data is sampled from the posterior predictive of the model. In principle we may want to consider sampling from the prior predictive.}

\item{disc_fun}{A function of the form \verb{function(mcmc_samples, ...)}
that computes and returns a scalar posterior predictive p-value (PPP) or discrepancy summary for that replicate.}

\item{num_reps}{Integer. Number of calibration replicates to run.}

\item{control}{Optional list of controls passed to \code{MCMC_fun}
(e.g., \code{niter}, seeds, etc.).}

\item{...}{Additional arguments passed through to the user-supplied functions.}
}
\description{
This function runs the calibration to obtain the calibrated posterior predictive p-value.
}
\details{
This function is fully engine-agnostic and contains no NIMBLE-specific code. It provides the scaffolding for the cppp calibration procedure.
}
