% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCPPP.R
\name{runCalibration}
\alias{runCalibration}
\title{Run calibration}
\usage{
runCalibration(
  data,
  nCalibrationReplicates = 100,
  MCMCcontrol = list(niter = 200, thin = 1, nburnin = 0),
  sampleFromPosterior,
  sampleFromPP,
  discrepancy,
  origMCMCSamples,
  ...
)
}
\arguments{
\item{data:}{original data}

\item{nCalibrationReplicates:}{number of calibration replicates}

\item{MCMCcontrol:}{list with number of posterior samples per replicates (and optional thinning and burnin)}

\item{sampleFromPosterior:}{function that can sample from the posterior distribution}

\item{sampleFromPP:}{function that can sample new data from the posterior predictive using some provided samples}

\item{discrepancy:}{a function that calculates the discrepancy (statistics) for the data (original or sampled)}

\item{origMCMCSamples:}{matrix of posterior samples from the original data}

\item{...:}{additional arguments to be passed}
}
\description{
This function runs the calibration (aka parametric bootstrap) to obtain the calibrated posterior predictive p-value.
}
