% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibration.R
\name{runCalibration}
\alias{runCalibration}
\title{Run posterior predictive calibration}
\usage{
runCalibration(
  MCMCSamples,
  observedData,
  MCMCFun,
  simulateNewDataFun,
  discFun,
  nReps,
  drawIndexSelector = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{MCMCSamples}{Matrix of posterior draws from the observed-data fit.}

\item{observedData}{Observed dataset (any R object).}

\item{MCMCFun}{Function \verb{function(newData, control)} that runs a short MCMC  on \code{newData} and returns posterior samples.}

\item{simulateNewDataFun}{Function \verb{function(thetaRow, observedData, control)} that  simulates one replicated dataset from the posterior predictive. SP: We assume that new data is sampled from the posterior predictive of the model. In principle we may want to consider sampling from the prior predictive.}

\item{discFun}{Function \verb{function(MCMCSamples, data, control)} that returns  a list with at least a numeric vector \code{D} of discrepancy values, one per row
of \code{MCMCSamples}.}

\item{nReps}{Number of calibration replications.}

\item{drawIndexSelector}{Optional function \verb{function(MCMCSamples, nReps, control)}
returning the indices of rows to use as seeds for calibration.}

\item{control}{List of additional backend-specific arguments.}

\item{...}{Not used currently.}
}
\value{
a list (future \code{S3} class \code{cpppResult} objects) containing the CPPP, observed and replicated repPPP, observed discrepancies and replicated discrepancies.
}
\description{
Run posterior predictive calibration
}
