% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCalibrationNIMBLE.R
\name{runCalibrationNIMBLE}
\alias{runCalibrationNIMBLE}
\title{Run calibration using a NIMBLE model}
\usage{
runCalibrationNIMBLE(
  model,
  dataNames = NULL,
  paramNames,
  discFun,
  simulateNewDataFun,
  nReps = 100,
  MCMCcontrolMain = list(niter = 5000, nburnin = 1000, thin = 1),
  MCMCcontrolRep = list(niter = 500, nburnin = 0, thin = 1),
  mcmcConfFun = NULL,
  drawIndexSelector = NULL,
  control = list(),
  ...
)
}
\arguments{
\item{model}{either an uncompiled or compiled nimbleModel with observed data set.}

\item{dataNames}{Optional character vector of data node names. If NULL,
nodes flagged as data in the model are used.}

\item{paramNames}{Character vector of parameter node names to monitor. (SP: if NULL some default?)}

\item{discFun}{Function \verb{function(MCMCSamples, targetData, control)} that returns \code{list(obs, sim)} with one discrepancy value per posterior draw of \code{MCMCSamples}.}

\item{simulateNewDataFun}{Function \verb{function(thetaRow, control)} that  simulates one replicated dataset from the posterior predictive. SP: We assume that new data is sampled from the posterior predictive of the model. In principle we may want to consider sampling from the prior predictive.}

\item{nReps}{Number of calibration replications.}

\item{MCMCcontrolMain}{List with \code{niter}, \code{nburnin}, \code{thin} for main chain.}

\item{MCMCcontrolRep}{List with \code{niter}, \code{nburnin}, \code{thin} for calibration chains.}

\item{mcmcConfFun}{Optional function \verb{function(model)} returning an MCMC configuration.}

\item{drawIndexSelector}{Optional row selector (see runCalibration()).}

\item{control}{List of additional options passed to discrepancy / helpers.}

\item{...}{Not used currently.}
}
\description{
Run calibration using a NIMBLE model
}
